workish

+++ install.sh
@@ -1,5 +1,18 @@
 #!/bin/bash
 
+DEVICE_FILE="devices.tmp"
+
+if [ ! -f "$DEVICE_FILE" ]; then
+    echo "Device file does not exist: $DEVICE_FILE"
+    exit 1
+fi
+
+# Check for environment variables or ask for them
+: "${USERNAME:=$(read -p "Enter the username: " REPLY; echo $REPLY)}"
+: "${PASSWORD:=$(read -p "Enter the password for $USERNAME: " REPLY; echo $REPLY)}"
+: "${HOSTNAME:=$(read -p "Enter the hostname: " REPLY; echo $REPLY)}"
+: "${GETSSHID:=$(read -p "Enter the ssh-import-id (i.e. gh:username for github): " REPLY; echo $REPLY)}"
+
 apt update -y
 apt upgrade -y
 apt install -y ubuntu-server
@@ -13,8 +26,11 @@ apt install -y linux-{,image-,headers-}generic linux-firmware \
                openssh-server \
                keyutils curl wget parted command-not-found ssh-import-id \
                mc
-ssh-import-id gh:killua-eu
-ssh-import-id gh:killua-eu -o /etc/dropbear/initramfs/authorized_keys
+
+ssh-import-id "${GETSSHID}"
+ssh-import-id "${GETSSHID}" | sudo tee /etc/dropbear-initramfs/authorized_keys > /dev/null
+sudo chmod 600 /etc/dropbear-initramfs/authorized_keys
+
 cd /etc/dropbear/initramfs
 
 echo "GRUB_ENABLE_CRYPTODISK=y" > /etc/default/grub.d/cryptodisk.cfg
@@ -34,28 +50,25 @@ echo "crypt2 UUID=$(blkid -s UUID -o value /dev/disk/by-partlabel/prim2 | tr -d 
 
 update-grub
 update-initramfs -u -k all
-grub-install /dev/vda # !!!!!!
-grub-install /dev/vdb # !!!!!!
+while IFS= read -r device; do
+    echo "Running grub-install on $device..."
+    grub-install "$device"
+done < "$DEVICE_FILE"
 grub-install --target=x86_64-efi --efi-directory=/boot/efi --bootloader-id=Ubuntu
 update-grub
 
+echo "${HOSTNAME}" > /etc/hostname
+sudo useradd ${USERNAME} -mG users,sudo,adm,plugdev,lxd -s /bin/bash
+echo "${USERNAME}:${PASSWORD}" | sudo chpasswd
 
-FIRSTUSER="killua"
-FIRSTPASS="killua"
-SETHOSTNAME="myhost"
-IMPORTSSH="gh:killua-eu"
-echo "${SETHOSTNAME}" > /etc/hostname
-sudo useradd ${FIRSTUSER} -mG users,sudo,adm,plugdev,lxd -s /bin/bash
-echo "${FIRSTUSER}:${FIRSTPASS}" | sudo chpasswd
-
-ssh-import-id $IMPORTSSH -o /home/${FIRSTUSER}/.ssh/authorized_keys
+ssh-import-id GETSSHID -o /home/${USERNAME}/.ssh/authorized_keys
 
 # After first boot
 timedatectl set-timezone Europe/Prague
 systemctl enable fstrim.timer
 systemctl enable systemd-timesyncd
 hostnamectl status
-hostnamectl set-hostname ${SETHOSTNAME}
+hostnamectl set-hostname ${HOSTNAME}
